fix(fr/raijinscans): port paginated reader, convert to standalone source
Site moved to an obfuscated AJAX reader manifest with cursor pagination
(per-load randomized field names via protocol request/response key maps).
Port haruneko eb7a0b2: rewrite pageListParse to extract the manifest and
loop nextOffset/nextToken/hasMore until every page is collected.

Also convert the extension from the lib-multisrc theme to a standalone
src/fr/raijinscans source and remove Pornhwa.io, the only other theme
consumer. extVersionCode = 59.
