--- lib-multisrc/raijinscans/build.gradle.kts
@@ -1,7 +1,0 @@
-import keiyoushi.gradle.extensions.baseVersionCode
-
-plugins {
-    alias(kei.plugins.multisrc)
-}
-
-baseVersionCode = 1
