--- src/fr/pornhwaio/build.gradle
@@ -1,10 +1,0 @@
-ext {
-    extName = 'Pornhwa.io'
-    extClass = '.PornhwaIo'
-    themePkg = 'raijinscans'
-    baseUrl = 'https://pornhwa.io'
-    overrideVersionCode = 0
-    isNsfw = true
-}
-
-apply plugin: "kei.plugins.extension.legacy"
