--- src/fr/pornhwaio/src/eu/kanade/tachiyomi/extension/fr/pornhwaio/PornhwaIo.kt
@@ -1,7 +1,0 @@
-package eu.kanade.tachiyomi.extension.fr.pornhwaio
-
-import eu.kanade.tachiyomi.multisrc.raijinscans.RaijinScans
-
-class PornhwaIo : RaijinScans("Pornhwa.io", "https://pornhwa.io", "fr") {
-    override val hasPremiumChapters = false
-}
